Edit Job Front, Add One Student Jobs, Payment Button

+++ server/routes/job.js
@@ -26,18 +26,34 @@ router.post('/', (req, res) => {
 })
 
 /*
-  [정상] 클래스 별 모든 Job 불러오기
-  : Job에서 classId 이용. {classId:}
+  [정상] 선생님용클래스 별 모든 Job 불러오기
+  : Job에서 classId 이용. {classId:} ondelete 관계 없이
 */
-router.get('/', (req, res) => {
+router.get('/manage', (req, res) => {
   //console.log('job get', req.query)
-  Job.find(req.query, function (err, jobs) {
+  const classId = req.query.classId
+  Job.find({ classId: classId }, function (err, jobs) {
     //console.log(jobs)
     const result = jobs
     if (err) return res.status(500).json({ error: err });
     res.json(result)
   })
 })
+/*
+  [정상] 학생용 클래스 별 모든 Job 불러오기 -apply하는 곳에서는 안보여야 한다
+  : Job에서 classId 이용. {classId:} ondelete:false
+*/
+router.get('/', (req, res) => {
+  //console.log('job get', req.query)
+  const classId = req.query.classId
+  Job.find({ classId: classId, ondelete: false }, function (err, jobs) {
+    //console.log(jobs)
+    const result = jobs
+    if (err) return res.status(500).json({ error: err });
+    res.json(result)
+  })
+})
+
 /*
   [정상] Job 정보 업데이트
   : Job 에서 Job._id 이용
@@ -51,25 +67,30 @@ router.put('/', (req, res) => {
   })
 })
 /*
-  [] Job 삭제 -> 예금 삭제처럼 로직 바꿔야 함.
-  : JoinedUser에 저장되어 있는 JobId도 없어져야 함.
+  [완료] Job 삭제 -> 예금 삭제처럼 로직됨
 */
 router.delete('/:id', async (req, res) => {
   const jobId = req.params.id;
-  console.log(jobId)
+
   const session = await startSession();
   try {
     session.startTransaction();
-    
-    //job 
-    const res1 = await JoinedUser.updateMany({}, { $pull: { jobId: jobId } }).exec({ session })
-    console.log(res1)
-    
-    
-    //job 삭제
-    const res2 = await Job.deleteOne({ _id: jobId }).exec({ session })
-    console.log(res2)
-    
+
+    // joinedUser에 jobId가 없다면 (해당 job을 가진 학생이 없다)
+    const havejob = await JoinedUser.countDocuments({ jobId: jobId }).exec({ session })
+    console.log(havejob)
+    if (havejob <= 0) {
+      //job 
+      const res1 = await JoinedUser.updateMany({}, { $pull: { jobId: jobId } }).exec({ session })
+      //console.log(res1)
+
+      //job 삭제
+      const res2 = await Job.deleteOne({ _id: jobId }).exec({ session })
+      //console.log(res2)
+    } else {
+      const res3 = await Job.updateOne({ _id: jobId }, { $set: { ondelete: true } }).exec({ session })
+    }
+
     // 트랜젝션 커밋
     await session.commitTransaction();
     // 트랜젝션 종료
@@ -85,37 +106,37 @@ router.delete('/:id', async (req, res) => {
 })
 
 /*
-  >>>>>>>>>>>>>해당 직업을 가지고 있는 모든 학생
+  [정상]해당 직업을 가지고 있는 모든 학생
 */
-router.get('/:id/students',(req,res)=>{
-  const jobId=req.params.id
-  const classId=req.query.classId
+router.get('/:id/students', (req, res) => {
+  const jobId = req.params.id
+  const classId = req.query.classId
   JoinedUser.aggregate([
     {
       $match: {
-        'classId':ObjectId(classId),
+        'classId': ObjectId(classId),
         'jobId': ObjectId(jobId)
       }
     },
     {
-      $project:{
-        "userId":'$userId',
+      $project: {
+        "userId": '$userId',
       }
     },
     {
-      $lookup:{
-        from :'users',
-        localField:'userId',
-        foreignField:'_id',
-        as:'user'
+      $lookup: {
+        from: 'users',
+        localField: 'userId',
+        foreignField: '_id',
+        as: 'user'
       }
     },
     {
-      $unwind:'$user'
+      $unwind: '$user'
     },
     {
-      $project:{
-        "name":'$user.name'
+      $project: {
+        "name": '$user.name'
       }
     },
   ]).exec((err, employee) => {
@@ -141,7 +162,8 @@ router.get('/:id/students',(req,res)=>{
   [완성]Job마다 salary 부여&세금 부여
 */
 router.post('/salary', async (req, res) => {
-  const classId = req.query.classId
+  console.log('/salary', req.body)
+  const classId = req.body.classId
   const session = await startSession();
   try {
     session.startTransaction();
@@ -150,17 +172,17 @@ router.post('/salary', async (req, res) => {
     const jobs = await JoinedUser.find({ classId: classId }).exec({ session })
     //console.log('jobs',jobs)
 
-    for await(const v of jobs) {
+    for await (const v of jobs) {
       const userjob = await Job.find({ '_id': { $in: v.jobId } }).exec({ session })
       //console.log('*')
-      for await(const job of userjob) {//user가 가지고 있는 job들
+      for await (const job of userjob) {//user가 가지고 있는 job들
         //console.log('x')
         const account = await Account.findOne({ studentId: v._id })
         //console.log(account.currentBalance,account.alias,'+',job.salary)
         // 2) 월급
         const give = await Account.updateOne({ _id: account._id }, { $inc: { currentBalance: job.salary } }).exec({ session })
         //console.log(give)
-        const after = await Account.findOne({studentId:v._id})
+        const after = await Account.findOne({ studentId: v._id })
         //console.log(after.currentBalance)
         const grantsalary = new AccountTransaction({
           accountId: account._id,
@@ -175,21 +197,22 @@ router.post('/salary', async (req, res) => {
         const tax = await Tax.findOne({ classId: classId }).exec({ session })
         const incometax = await Math.round(tax.taxlist.income * job.salary / 100)
         //3-2) 통장에서 세금 징수
-        const minus = await Account.updateOne({ _id: account._id }, { $inc: { currentBalance: - incometax } }).exec({ session })
-        const after2 = await Account.findOne({studentId:v._id})
-        //console.log('-',incometax,after2.currentBalance)
-        //3-3) 통장에 내역 기록 
-        const paytax = new AccountTransaction({
-          accountId: account._id,
-          transactionType: 0,
-          amount: incometax,
-          afterbalance: after2.currentBalance,//
-          memo: '소득세'
-        })
-        await paytax.save({ session })
-        // 3-4) Budget의 income에 추가
-        const budget=await Budget.updateOne({classId:classId},{$inc: { "balance.income": + incometax }}).exec({session})
-      
+        if (incomeetax > 0) {
+          const minus = await Account.updateOne({ _id: account._id }, { $inc: { currentBalance: - incometax } }).exec({ session })
+          const after2 = await Account.findOne({ studentId: v._id })
+          //console.log('-',incometax,after2.currentBalance)
+          //3-3) 통장에 내역 기록 
+          const paytax = new AccountTransaction({
+            accountId: account._id,
+            transactionType: 0,
+            amount: incometax,
+            afterbalance: after2.currentBalance,//
+            memo: '소득세'
+          })
+          await paytax.save({ session })
+          // 3-4) Budget의 income에 추가
+          const budget = await Budget.updateOne({ classId: classId }, { $inc: { "balance.income": + incometax } }).exec({ session })
+        }
       }
     }
